requirements.txt
numpy
opencv-contrib-python
paho-mqtt
